+++ src/library/DIPS.Mobile.UI/Components/Images/ImageButton/ImageButtonHandler.cs
@@ -8,8 +8,11 @@ public partial class ImageButtonHandler : Microsoft.Maui.Handlers.ImageButtonHan
     
     public static IPropertyMapper<ImageButton, ImageButtonHandler> PropertyMapper = new PropertyMapper<ImageButton, ImageButtonHandler>(Mapper)
     {
-        [nameof(ImageButton.TintColor)] = TrySetTintColor
+        [nameof(ImageButton.TintColor)] = TrySetTintColor,
+        [nameof(ImageButton.AdditionalHitBoxSize)] = MapAdditionalHitBoxSize
     };
+
+    private static partial void MapAdditionalHitBoxSize(ImageButtonHandler handler, ImageButton imageButton);
 
     private static partial void TrySetTintColor(ImageButtonHandler handler, ImageButton imageButton);