+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## [12.11.0]
+- Extended dui:Button and dui:ImageButton so consumers can set additional hitbox size.
+- Added ripple effect to dui:ImageButton.
 
 ## [12.10.0]
 - Added DateTimeFormatter